second overflow fix

+++ trunk/avidemux/avidemux/oplug_mpeg/op_mpegwrite.cpp
@@ -760,7 +760,7 @@ uint32_t		len,flags,type,outquant,audiolen;
 	mpegvbr.fps=_fps1000/1000.;
 
 	mpegvbr.mode=VBR_MODE_2PASS_2;
-	mpegvbr.desired_size=final_size*1024*1024;
+	mpegvbr.desired_size=(uint64_t)final_size*(uint64_t)1024*(uint64_t)1024;
 	mpegvbr.debug=0;
 	mpegvbr.filename=statname; //XvidInternal2pass_statfile;
